Function rb_sys::bindings::rb_Rational

pub unsafe extern "C" fn rb_Rational(num: VALUE, den: VALUE) -> VALUE

Converts various values into a Rational. This function accepts:

Instances of ::rb_cInteger (taken as-is),

Instances of ::rb_cRational (taken as-is),

Instances of ::rb_cFloat (applies #to_r),

Instances of ::rb_cComplex (applies #to_r),

Instances of ::rb_cString (applies #to_r),

Other objects that respond to #to_r.

It (possibly recursively) applies #to_r until both sides become either Integer or Rational, then divides them.

As a special case, passing ::RUBY_Qundef to den is the same as passing RB_INT2NUM(1).

@param in num Numerator (see above).

@param in den Denominator (see above).

@exception rb_eTypeError Passed something not described above.

@exception rb_eFloatDomainError #to_r produced Nan/Inf.

@exception rb_eZeroDivError #to_r produced zero for den.

@return An instance of ::rb_cRational whose value is (num/den)r.

Internal

This was the implementation of Kernel#Rational before, but they diverged.

Generated by rb-sys for Ruby 3.2.0