Function rb_sys::bindings::rb_ary_plus

pub unsafe extern "C" fn rb_ary_plus(lhs: VALUE, rhs: VALUE) -> VALUE

Creates a new array, concatenating the former to the latter.

@param in lhs Source array #1.

@param in rhs Source array #2.

@exception rb_eIndexError Result array too big.

@return A new array containing rhs concatenated to lhs.

Note

This operation doesn’t commute. Don’t get confused by the "plus" terminology. For historical reasons there are some noncommutative +s in Ruby. This is one of such things. There has been a long discussion around +s in programming languages.

Internal

rb_ary_concat is not a destructive version of rb_ary_plus. They raise different exceptions. Don’t know why though.

Generated by rb-sys for Ruby 3.2.0