Function rb_sys::bindings::rb_check_id

pub unsafe extern "C" fn rb_check_id(namep: *mut VALUE) -> ID

Detects if the given name is already interned or not. It first tries to convert the argument to an instance of ::rb_cString if it is neither an instance of ::rb_cString nor ::rb_cSymbol. The conversion result is written back to the variable. Then queries if that name was already interned before. If found it returns such id, otherwise zero.

We eventually introduced this API to avoid inadvertent symbol pin-down. Before, there was no way to know if an ID was already interned or not without actually creating one (== leaking memory). By using this API you can avoid such situations:

bool does_interning_this_leak_memory(VALUE obj)
{
auto tmp = obj;
if (auto id = [`rb_check_id`]&tmp); id) {
return false;
}
else {
return true; // Let GC sweep tmp if necessary.
}
}

@param in,out namep A pointer to a name to query.

@pre The object referred by *namep must either be an instance of ::rb_cSymbol, or an instance of ::rb_cString, or responds to #to_str method.

@exception rb_eTypeError Can’t convert *namep into ::rb_cString.

@exception rb_eEncodingError Given string is non-ASCII.

@retval 0 No such id ever existed in the history.

@retval otherwise The id that represents the given name.

@post The object that *namep points to is a converted result object, which is always an instance of either ::rb_cSymbol or ::rb_cString.

@see https://bugs.ruby-lang.org/issues/5072

Internal

@shyouhei doesn’t know why this has to raise rb_eEncodingError.

Generated by rb-sys for Ruby 3.2.0