Function rb_sys::bindings::rb_close_before_exec

pub unsafe extern "C" fn rb_close_before_exec(
    lowfd: c_int,
    maxhint: c_int,
    noclose_fds: VALUE
)

Closes everything. In case of POSIX environments, a child process inherits its parent’s opened file descriptors. Which is nowadays considered as one of the UNIX mistakes. This function closes such inherited file descriptors. When your C extension needs to have a child process, don’t forget to call this from your child process right before exec.

@param in lowfd Lower bound of FDs (you want STDIN to remain, no?).

@param in maxhint Hint of max FDs.

@param in noclose_fds A hash, whose keys are an allowlist.

Internal

As of writing, in spite of the name, this function does not actually close anything. It just sets FD_CLOEXEC for everything and let execve(2) to atomically close them at once. This is because as far as we know there are no such platform that has fork(2) but lacks FD_CLOEXEC.

Because this function is expected to run on a forked process it is entirely async-signal-safe.

Generated by rb-sys for Ruby 3.2.0