Function rb_sys::bindings::rb_define_class

pub unsafe extern "C" fn rb_define_class(
    name: *const c_char,
    super_: VALUE
) -> VALUE

Defines a top-level class.

@param in name Name of the class.

@param in super A class from which the new class will derive.

@exception rb_eTypeError The constant name name is already taken but the constant is not a class.

@exception rb_eTypeError The class is already defined but the class can not be reopened because its superclass is not super.

@exception rb_eArgError super is NULL.

@return The created class.

@post Top-level constant named name refers the returned class.

Note

If a class named name is already defined and its superclass is super, the function just returns the defined class.

Note

The compaction GC does not move classes returned by this function.

Internal

There are classes without names, but you can’t pass NULL here. You have to use other ways to create one.

Generated by rb-sys for Ruby 3.2.0