Function rb_sys::bindings::rb_econv_has_convpath_p

pub unsafe extern "C" fn rb_econv_has_convpath_p(
    from_encoding: *const c_char,
    to_encoding: *const c_char
) -> c_int

Queries if there is more than one way to convert between the passed two encodings. Encoding conversion are has_and_belongs_to_many relationships. There could be no direct conversion defined for the passed pair. Ruby tries to find an indirect way to do so then. For instance ISO-8859-1 has no direct conversion to ISO-2022-JP. But there is ISO-8859-1 to UTF-8 conversion; then there is UTF-8 to EUC-JP conversion; finally there also is EUC-JP to ISO-2022-JP conversion. So in short ISO-8859-1 can be converted to ISO-2022-JP using that path. This function returns true. Obviously not everything that can be represented using UTF-8 can also be represented using EUC-JP. Conversions in practice can fail depending on the actual input, and that renders exceptions in case of rb_str_encode.

@param in from_encoding One encoding.

@param in to_encoding Another encoding.

@retval 0 No way to convert the two.

@retval 1 At least one way to convert the two.

Internal

Practically @shyouhei knows no way for this function to return 0. It seems everything can eventually be converted to/from UTF-8, which connects everything.

Generated by rb-sys for Ruby 3.2.0