Function rb_sys::bindings::rb_exit

pub unsafe extern "C" fn rb_exit(status: c_int)

Terminates the current execution context. This API is the entry point of a "well-mannered" termination sequence. When called from an extension library, it raises ::rb_eSystemExit exception. Programs could rescue that exception. Can cancel process exit then. Otherwise, that exception results in a process termination with the status passed to this function.

@param in status Exit status, see also exit(3).

@exception rb_eSystemExit Exception representing the exit status.

Note

It never returns.

Internal

"When called from an extension library"? You might wonder. In fact there are chances for this function to be called from outside of it, for instance when dlopen(3) failed. In case it is not possible for this function to raise an exception, it does not (silently enters to process cleanup). But that is a kind of implementation detail which extension library authors should not bother.

Generated by rb-sys for Ruby 3.2.0