Function rb_sys::bindings::rb_fiber_resume

pub unsafe extern "C" fn rb_fiber_resume(
    fiber: VALUE,
    argc: c_int,
    argv: *const VALUE
) -> VALUE

Resumes the execution of the passed fiber, either from the point at which the last rb_fiber_yield was called if any, or at the beginning of the fiber body if it is the first call to this function.

Other arguments are passed into the fiber’s body, either as return values of rb_fiber_yield in case it switches to there, or as the block parameter of the fiber body if it switches to the beginning of the fiber.

The return value of this function is either the value passed to previous rb_fiber_yield call, or the ultimate evaluated value of the entire fiber body if the execution reaches the end of it.

When an exception happens inside of a fiber it propagates to this function.

f = Fiber.new do |i|
puts \"<x> =>> #{i}\
puts \"<y> <-- #{i + 1}\
j = Fiber.yield(i + 1)
puts \"<z> =>> #{j}\
puts \"<w> <-- #{j + 1}\
next j + 1
end

puts \"[a] <-- 1\
p = f.resume(1)
puts \"[b] =>> #{p}\
puts \"[c] <-- #{p + 1}\
q = f.resume(p + 1)
puts \"[d] =>> #{q}\

Above program executes in [a] <x> <y> [b] [c] <z> <w> [d].

@param out fiber The fiber to resume.

@param in argc Number of objects of argv.

@param in argv Passed (somehow) to fiber.

@exception [rb_eFiberError] fib is terminated etc.

@exception rb_eException Any exceptions happen in fiber.

@return (See above)

Note

This function does return.

Internal

@shyouhei expected this function to raise ::rb_eFrozenError for frozen fibers but it doesn’t in practice. Intentional or …?

Generated by rb-sys for Ruby 3.2.0