Function rb_sys::bindings::rb_gc_mark_movable

pub unsafe extern "C" fn rb_gc_mark_movable(obj: VALUE)

Maybe this is the only function provided for C extensions to control the pinning of objects, so let us describe it in detail. These days Ruby’s GC is copying. As far as an object’s physical address is guaranteed unused, it can move around the object space. Our GC engine rearranges these objects after it reclaims unreachable objects from our object space, so that the space is compact (improves memory locality). This is called the "compaction" phase, and works well most of the time… as far as there are no C extensions. C extensions complicate the scenario because Ruby core cannot detect any use of the physical address of an object inside of C functions. In order to prevent memory corruptions, objects observable from C extensions are "pinned"; they stick to where they are born until they die, just in case any C extensions touch their raw pointers. This variant of scheme is called "Mostly-Copying" garbage collector. Authors of C extensions, however, can extremely carefully write them to become compaction-aware. To do so avoid referring to a Ruby object from inside of your struct in the first place. But if that is not possible, use this function from your ::rb_data_type_struct::dmark then. This way objects marked using it are considered movable. If you chose this way you have to manually fix up locations of such moved pointers using rb_gc_location.

@see Bartlett, Joel F., "Compacting Garbage Collection with Ambiguous Roots", ACM SIGPLAN Lisp Pointers Volume 1 Issue 6 pp. 3-12, April-May-June, 1988. https://doi.org/10.1145/1317224.1317225

@param in obj Object that is movable.

@post Values stored in tbl are marked.

Generated by rb-sys for Ruby 3.2.0