Function rb_sys::bindings::rb_io_stdio_file

pub unsafe extern "C" fn rb_io_stdio_file(
    fptr: *mut rb_io_t
) -> *mut FILE

Finds or creates a stdio’s file structure from a Ruby’s one. This can be handy if you want to call an external API that accepts FILE *.

Note

Note however, that FILEs can have their own buffer. Mixing Ruby’s and stdio’s file are basically dangerous. Use with care.

@param in,out fptr Target IO.

@return A stdio’s file, created if absent.

@post fptr has its corresponding stdio’s file.

Internal

We had rich support for FILE before! In the days of 1.8.x ::rb_io_t was like this:

typedef struct [`rb_io_t`] {
FILE *f;                    // stdio ptr for read/write
FILE *f2;                   // additional ptr for rw pipes
int mode;                   // mode flags
int pid;                    // child's pid (for pipes)
int lineno;                 // number of lines read
char *path;                 // pathname for file
void (*finalize) _((struct [`rb_io_t`]*,int)); // finalize proc
} [`rb_io_t`];

But we eventually abandoned this layout. It was too difficult. We could not have fine-grained control over the f field.

FILE tends to be an opaque struct. It does not interface well with select(2) etc. This makes IO multiplexing quite hard. Using stdio, there is arguably no portable way to know if fwrite(3) blocks.

Nonblocking mode, which is another core concept that enables IO multiplexing, does not interface with stdio routines at all.

Detection of duplexed IO is also hard for the same reason.

feof(3) is not portable. https://mail.python.org/pipermail/python-dev/2001-January/011390.html

Solaris was a thing back then. They could not have more than 256 FILE structures at a time. Their file descriptors ware stored in an unsigned char.

It is next to impossible to avoid SEGV, especially when a thread tries to ungetc(3)-ing from a FILE which is fread(3)-ed by another one.

In short, it is a bad idea to let someone else manage IO buffers, especially someone you cannot control. This still applies to extension libraries methinks. Ruby doesn’t prevent you from shooting yourself in the foot, but consider yourself warned here.

Generated by rb-sys for Ruby 3.2.0