Function rb_sys::bindings::rb_marshal_define_compat

pub unsafe extern "C" fn rb_marshal_define_compat(
    newclass: VALUE,
    oldclass: VALUE,
    dumper: Option<unsafe extern "C" fn(arg1: VALUE) -> VALUE>,
    loader: Option<unsafe extern "C" fn(arg1: VALUE, arg2: VALUE) -> VALUE>
)

Marshal format compatibility layer. Over time, classes evolve, so that their internal data structure change drastically. For instance an instance of ::rb_cRange was made of ::RUBY_T_OBJECT in 1.x., but in 3.x it is a ::RUBY_T_STRUCT now. In order to keep binary compatibility, we "fake" the marshalled representation to stick to old types. This is the API to enable that manoeuvre. Here is how:

First, because you are going to keep backwards compatibility, you need to retain the old implementation of your class. Rename it, and keep the class somewhere (for instance [rb_register_global_address] could help). Next create your new class. Do whatever you want.

Then, this is the key point. Create two new "bridge" functions that convert the structs back and forth:

the "dumper" function that takes an instance of the new class, and returns an instance of the old one. This is called from rb_marshal_dump, to keep it possible for old programs to read your new data.

the "loader" function that takes two arguments, new one and old one, in that order. rb_marshal_load calls this function when it finds a representation of the retained old class. The old one passed to this function is the reconstructed instance of the old class. Reverse-engineer that to modify the new one, to have the identical contents.

Finally, connect all of them using this function.

@param in newclass The class that needs conversion.

@param in oldclass Old implementation of newclass.

@param in dumper Function that converts newclass to oldclass.

@param in loader Function that converts oldclass to newclass.

@exception rb_eTypeError newclass has no allocator.

Generated by rb-sys for Ruby 3.2.0