Function rb_sys::bindings::rb_marshal_load

pub unsafe extern "C" fn rb_marshal_load(port: VALUE) -> VALUE

Deserialises a previous output of rb_marshal_dump into a network of objects.

@param in,out port Either IO or String.

@exception rb_eTypeError port is in unexpected type.

@exception rb_eArgError Contents of port is broken.

@return Object(s) rebuilt using the info from port.

SECURITY CONSIDERATIONS

Warning

By design, rb_marshal_load can deserialise almost any class loaded into the Ruby process. In many cases this can lead to remote code execution if the Marshal data is loaded from an untrusted source.

Warning

As a result, rb_marshal_load is not suitable as a general purpose serialisation format and you should never unmarshal user supplied input or other untrusted data.

Warning

If you need to deserialise untrusted data, use JSON or another serialisation format that is only able to load simple, ‘primitive’ types such as String, Array, Hash, etc. Never allow user input to specify arbitrary types to deserialise into.

Generated by rb-sys for Ruby 3.2.0