Function rb_sys::bindings::rb_mutex_sleep

pub unsafe extern "C" fn rb_mutex_sleep(
    self_: VALUE,
    timeout: VALUE
) -> VALUE

Releases the lock held in the mutex and waits for the period of time; reacquires the lock on wakeup.

@pre The lock has to be owned by the current thread beforehand.

@param out self The target mutex.

@param in timeout Duration, in seconds, in ::rb_cNumeric.

@exception rb_eArgError timeout is negative.

@exception rb_eRangeError timeout is out of range of time_t.

@exception rb_eThreadError The mutex is not owned by the current thread.

@return Number of seconds it actually slept.

Warning

It is a failure not to check the return value. This function can return spuriously for various reasons. Maybe other threads can rb_thread_wakeup. Maybe an end user can press the Control and C key from the interactive console. On the other hand it can also take longer than the specified. The mutex could be locked by someone else. It waits then.

@post Upon successful return the passed mutex is owned by the current thread.

Internal

This function is called from ConditionVariable#wait. So it is not a deprecated feature. However @shyouhei have never seen any similar mutex primitive available in any other languages than Ruby.

EDIT: In 2021, @shyouhei asked @ko1 in person about this API. He answered that it is his invention. The motivation behind its design is to eliminate needs of condition variables as primitives. Unlike other languages, Ruby’s ConditionVariable class was written in pure-Ruby initially. We don’t have to implement machine-native condition variables in assembly each time we port Ruby to a new architecture. This function made it possible. "I felt I was a genius when this idea came to me", said @ko1.

[rb_cConditionVariable] is now written in C for speed, though.

Generated by rb-sys for Ruby 3.2.0