Function rb_sys::bindings::rb_scan_args

pub unsafe extern "C" fn rb_scan_args(
    argc: c_int,
    argv: *const VALUE,
    fmt: *const c_char,
     ...
) -> c_int

Retrieves argument from argc and argv to given ::VALUE references according to the format string. The format can be described in ABNF as follows:

scan-arg-spec  := param-arg-spec [keyword-arg-spec] [block-arg-spec]

param-arg-spec        := pre-arg-spec [post-arg-spec] / post-arg-spec /
pre-opt-post-arg-spec
pre-arg-spec          := num-of-leading-mandatory-args
[num-of-optional-args]
post-arg-spec         := sym-for-variable-length-args
[num-of-trailing-mandatory-args]
pre-opt-post-arg-spec := num-of-leading-mandatory-args num-of-optional-args
num-of-trailing-mandatory-args
keyword-arg-spec      := sym-for-keyword-arg
block-arg-spec        := sym-for-block-arg

num-of-leading-mandatory-args  := DIGIT ; The number of leading mandatory
; arguments
num-of-optional-args           := DIGIT ; The number of optional arguments
sym-for-variable-length-args   := \"*\"   ; Indicates that variable length
;  arguments are captured as a ruby
; array
num-of-trailing-mandatory-args := DIGIT ; The number of trailing mandatory
; arguments
sym-for-keyword-arg            := \":\"   ; Indicates that keyword argument
; captured as a hash.
; If keyword arguments are not
; provided, returns nil.
sym-for-block-arg              := \"&\"   ; Indicates that an iterator block
; should be captured if given

For example, "12" means that the method requires at least one argument, and at most receives three (1+2) arguments. So, the format string must be followed by three variable references, which are to be assigned to captured arguments. For omitted arguments, variables are set to ::RUBY_Qnil. NULL can be put in place of a variable reference, which means the corresponding captured argument(s) should be just dropped.

The number of given arguments, excluding an option hash or iterator block, is returned.

@param in argc Length of argv.

@param in argv Pointer to the arguments to parse.

@param in fmt Format, in the language described above.

@param out … Variables to fill in.

@exception rb_eFatal Malformed fmt.

@exception rb_eArgError Arity mismatch.

@return Actually parsed number of given arguments.

@post Each values passed to argv is filled into the variadic arguments, according to the format.

Generated by rb-sys for Ruby 3.2.0