Function rb_sys::bindings::rb_str_succ

pub unsafe extern "C" fn rb_str_succ(orig: VALUE) -> VALUE

Searches for the "successor" of a string. This function is complicated! This is the only function in the entire ruby API (either C or Ruby) that generates a string out of thin air. First, the successor to an empty string is a new empty string:

''.succ # => \"\

Otherwise the successor is calculated by "incrementing" characters. The first character to be incremented is the rightmost alphanumeric: or, if no alphanumerics, the rightmost character:

'THX1138'.succ # => \"THX1139\
'<<koala>>'.succ # => \"<<koalb>>\
'***'.succ # => '**+'

The successor to a digit is another digit, "carrying" to the next-left character for a "rollover" from 9 to 0, and prepending another digit if necessary:

'00'.succ # => \"01\
'09'.succ # => \"10\
'99'.succ # => \"100\
'-9'.succ # => \"-10\

The successor to a letter is another letter of the same case, carrying to the next-left character for a rollover, and prepending another same-case letter if necessary:

'aa'.succ # => \"ab\
'az'.succ # => \"ba\
'zz'.succ # => \"aaa\
'AA'.succ # => \"AB\
'AZ'.succ # => \"BA\
'ZZ'.succ # => \"AAA\

The successor to a non-alphanumeric character is the next character in the underlying character set’s collating sequence, carrying to the next-left character for a rollover, and prepending another character if necessary:

s = \"\\u03A1\
s.succ # => \"\\u03A3\"  # There is no such thing like \\u03A2.
s = 255.chr * 3
s # => \"\\xFF\\xFF\\xFF\
s.succ # => \"\\x01\\x00\\x00\\x00\

Carrying can occur between and among mixtures of alphanumeric characters:

s = 'zz99zz99'
s.succ # => \"aaa00aa00\
s = '99zz99zz'
s.succ # => \"100aa00aa\
s = '1.9.9'
s.succ # => \"2.0.0\

@param in orig Predecessor string.

@return Successor string.

Generated by rb-sys for Ruby 3.2.0