Function rb_sys::bindings::rb_thread_call_without_gvl

pub unsafe extern "C" fn rb_thread_call_without_gvl(
    func: Option<unsafe extern "C" fn(arg1: *mut c_void) -> *mut c_void>,
    data1: *mut c_void,
    ubf: rb_unblock_function_t,
    data2: *mut c_void
) -> *mut c_void

Allows the passed function to run in parallel with other Ruby threads.

What this function does:

Checks (and handles) pending interrupts.

Releases the GVL. (Others can run here in parallel…)

Calls the passed function.

Blocks until it re-acquires the GVL.

Checks interrupts that happened between 2 to 4.

In case other threads interfaced with this thread using rb_thread_kill etc., the passed UBF is additionally called. See ::rb_unblock_function_t for details.

Unlike rb_thread_call_without_gvl2 this function also reacts to signals etc.

@param in func A function to call without GVL.

@param in,out data1 Passed as-is to func.

@param in ubf An UBF to cancel func.

@param in,out data2 Passed as-is to ubf.

@return What func returned, or 0 in case ubf cancelled func.

Warning

You cannot use most of Ruby C APIs like calling methods or raising exceptions from any of the functions passed to it. If that is dead necessary use rb_thread_call_with_gvl to re-acquire the GVL.

Warning

In short, this API is difficult. @ko1 recommends you to use other ways if any. We lack experiences to use this API. If you find any corner cases etc., please report it to the devs.

Warning

Releasing and re-acquiring the GVL are expensive operations. For a short-running func, it might be faster to just call func with blocking everything else. Be sure to benchmark your code to see if it is actually worth releasing the GVL.

Generated by rb-sys for Ruby 3.2.0