Function rb_sys::bindings::rb_tracepoint_new

pub unsafe extern "C" fn rb_tracepoint_new(
    target_thread_not_supported_yet: VALUE,
    events: rb_event_flag_t,
    func: Option<unsafe extern "C" fn(arg1: VALUE, arg2: *mut c_void)>,
    data: *mut c_void
) -> VALUE

Creates a tracepoint by registering a callback function for one or more tracepoint events. Once the tracepoint is created, you can use rb_tracepoint_enable to enable the tracepoint.

@param in target_thread_not_supported_yet Meant for picking the thread in which the tracepoint is to be created. However, current implementation ignore this parameter, tracepoint is created for all threads. Simply specify Qnil.

@param in events Event(s) to listen to.

@param in func A callback function.

@param in,out data Void pointer that will be passed to the callback function.

When the callback function is called, it will be passed 2 parameters:

VALUE tpval - the TracePoint object from which trace args can be extracted.

void *data - A void pointer which helps to share scope with the callback function.

It is important to note that you cannot register callbacks for normal events and internal events simultaneously because they are different purpose. You can use any Ruby APIs (calling methods and so on) on normal event hooks. However, in internal events, you can not use any Ruby APIs (even object creations). This is why we can’t specify internal events by TracePoint directly. Limitations are MRI version specific.

Example:

[`rb_tracepoint_new`]
Qnil,
RUBY_INTERNAL_EVENT_NEWOBJ | RUBY_INTERNAL_EVENT_FREEOBJ,
obj_event_i,
data);

In this example, a callback function obj_event_i will be registered for internal events #RUBY_INTERNAL_EVENT_NEWOBJ and #RUBY_INTERNAL_EVENT_FREEOBJ.

Generated by rb-sys for Ruby 3.2.0