Function rb_sys::bindings::rb_waitpid

pub unsafe extern "C" fn rb_waitpid(
    pid: pid_t,
    status: *mut c_int,
    flags: c_int
) -> pid_t

Waits for a process, with releasing GVL.

@param in pid Process ID.

@param out status The wait status is filled back.

@param in flags Wait options.

@retval -1 System call failed, errno set.

@retval 0 WNOHANG but no waitable children.

@retval otherwise A process ID that was wait()-ed.

@post Upon successful return status is updated to have the process’ status.

Note

status can be NULL.

Note

The arguments are passed through to underlying system call(s). Can have special meanings. For instance passing ([rb_pid_t]-1 to pid means it waits for any processes, under POSIX-compliant situations.

Generated by rb-sys for Ruby 3.2.0