Function rb_sys::bindings::ruby_vm_at_exit

pub unsafe extern "C" fn ruby_vm_at_exit(
    func: Option<unsafe extern "C" fn(arg1: *mut ruby_vm_t)>
)

ruby_vm_at_exit registers a function func to be invoked when a VM passed away. Functions registered this way runs in reverse order of registration, just like END {} block does. The difference is its timing to be triggered. ruby_vm_at_exit functions runs when a VM passed away, while END {} blocks runs just before a VM is passing away.

You cannot register a function to another VM than where you are in. So where to register is intuitive, omitted. OTOH the argument func cannot know which VM it is in because at the time of invocation, the VM has already died and there is no execution context. The VM itself is passed as the first argument to it.

@param in func the function to register.

Generated by rb-sys for Ruby 3.2.0