Struct rb_sys::bindings::rb_memory_view_t

#[repr(C)]
pub struct rb_memory_view_t {

    pub obj: VALUE,
    pub data: *mut c_void,
    pub byte_size: ssize_t,
    pub readonly: bool,
    pub format: *const c_char,
    pub item_size: ssize_t,
    pub item_desc: rb_memory_view_t__bindgen_ty_1,
    pub ndim: ssize_t,
    pub shape: *const ssize_t,
    pub strides: *const ssize_t,
    pub sub_offsets: *const ssize_t,
    pub private_data: *mut c_void,
    pub _memory_view_entry: *const rb_memory_view_entry,

}

A MemoryView structure, [rb_memory_view_t], is used for exporting objects’ MemoryView.

This structure contains the reference of the object, which is the owner of the MemoryView, the pointer to the head of exported memory, and the metadata that describes the structure of the memory. The metadata can describe multidimensional arrays with strides.

Generated by rb-sys for Ruby 3.2.0

Fields
obj: VALUE

The original object that has the memory exported via this memory view.

Generated by rb-sys for Ruby 3.2.0

data: *mut c_void

The pointer to the exported memory.

Generated by rb-sys for Ruby 3.2.0

byte_size: ssize_t

The number of bytes in data.

Generated by rb-sys for Ruby 3.2.0

readonly: bool

true for readonly memory, false for writable memory.

Generated by rb-sys for Ruby 3.2.0

format: *const c_char

A string to describe the format of an element, or NULL for unsigned bytes. The format string is a sequence of the following pack-template specifiers:

c, C, s, s!, S, S!, n, v, i, i!, I, I!, l, l!, L, L!, N, V, f, e, g, q, q!, Q, Q!, d, E, G, j, J, x

For example, "dd" for an element that consists of two double values, and "CCC" for an element that consists of three bytes, such as an RGB color triplet.

Also, the value endianness can be explicitly specified by ‘<’ or ‘>’ following a value type specifier.

The items are packed contiguously. When you emulate the alignment of structure members, put ‘|’ at the beginning of the format string, like "|iqc". On x86_64 Linux ABI, the size of the item by this format is 24 bytes instead of 13 bytes.

Generated by rb-sys for Ruby 3.2.0

item_size: ssize_t

The number of bytes in each element. item_size should equal to rb_memory_view_item_size_from_formatformat).

Generated by rb-sys for Ruby 3.2.0

item_desc: rb_memory_view_t__bindgen_ty_1

Generated by rb-sys for Ruby 3.2.0

ndim: ssize_t

The number of dimension.

Generated by rb-sys for Ruby 3.2.0

shape: *const ssize_t

ndim size array indicating the number of elements in each dimension. This can be NULL when ndim == 1.

Generated by rb-sys for Ruby 3.2.0

strides: *const ssize_t

ndim size array indicating the number of bytes to skip to go to the next element in each dimension.

Generated by rb-sys for Ruby 3.2.0

sub_offsets: *const ssize_t

The offset in each dimension when this memory view exposes a nested array. Or, NULL when this memory view exposes a flat array.

Generated by rb-sys for Ruby 3.2.0

private_data: *mut c_void

The private data for managing this exported memory

Generated by rb-sys for Ruby 3.2.0

_memory_view_entry: *const rb_memory_view_entry

DO NOT TOUCH THIS: The memory view entry for the internal use

Generated by rb-sys for Ruby 3.2.0